functions/requirements.txt
firebase_functions~=0.1.0
flask~=2.3.2
python-telegram-bot~=20.7
python-dotenv~=1.0.0
requests~=2.31.0
uvicorn~=0.24.0
asgiref~=3.7.2